fps fix
is nu ook speelbaar op mijn craptop

+++ Code/dirt.cpp
@@ -24,7 +24,7 @@ void dirt::seed()
         texture.loadFromImage(image);
         sf::Time time = clock.getElapsedTime();
         int64_t elpasedTime = time.asMicroseconds();
-        particles.generate(sf::Vector2i (position.x+5,position.y-5), elpasedTime, {128, 55, 15}, sf::Vector2f(6,6));
+        particles.generate(sf::Vector2f (position.x+5,position.y-5), elpasedTime, {128, 55, 15}, sf::Vector2f(6,6));
     }
 }
 
@@ -86,7 +86,7 @@ bool dirt::harvest()
         texture.loadFromImage(image);
         sf::Time time = clock.getElapsedTime();
         int64_t elpasedTime = time.asMicroseconds();
-        particles.generate(sf::Vector2i (position.x+5,position.y-5), elpasedTime, {237, 193, 33, 150}, sf::Vector2f(4,4));
+        particles.generate(sf::Vector2f (position.x+5,position.y-5), elpasedTime, {237, 193, 33, 150}, sf::Vector2f(4,4));
         ticks = 0;
         return true;
     }
